Move Custom type to index
