+++ app/src/ui/diff/text-diff.tsx
@@ -57,14 +57,15 @@ import * as OcticonSymbol from '../octicons/octicons.generated'
 import { WhitespaceHintPopover } from './whitespace-hint-popover'
 import { PopoverAnchorPosition } from '../lib/popover'
 import { DiffContentsWarning } from './diff-contents-warning'
+import { CustomOcticonSymbolType } from '../octicons'
 
 // This is a custom version of the no-newline octicon that's exactly as
 // tall as it needs to be (8px) which helps with aligning it on the line.
-export const narrowNoNewlineSymbol = {
+export const narrowNoNewlineSymbol: CustomOcticonSymbolType = {
   s: 'narrow-no-newline',
   w: 16,
   h: 8,
-  p: [
+  d: [
     'm 16,1 0,3 c 0,0.55 -0.45,1 -1,1 l -3,0 0,2 -3,-3 3,-3 0,2 2,0 0,-2 2,0 z M 8,4 C 8,6.2 6.2,8 4,8 1.8,8 0,6.2 0,4 0,1.8 1.8,0 4,0 6.2,0 8,1.8 8,4 Z M 1.5,5.66 5.66,1.5 C 5.18,1.19 4.61,1 4,1 2.34,1 1,2.34 1,4 1,4.61 1.19,5.17 1.5,5.66 Z M 7,4 C 7,3.39 6.81,2.83 6.5,2.34 L 2.34,6.5 C 2.82,6.81 3.39,7 4,7 5.66,7 7,5.66 7,4 Z',
   ],
 }
@@ -104,7 +105,7 @@ function createNoNewlineIndicatorWidget() {
   const widget = document.createElement('span')
   const titleId = uuid()
 
-  const { w, h, p } = narrowNoNewlineSymbol
+  const { w, h, d } = narrowNoNewlineSymbol
 
   const xmlns = 'http://www.w3.org/2000/svg'
   const svgElem = document.createElementNS(xmlns, 'svg')
@@ -122,7 +123,7 @@ function createNoNewlineIndicatorWidget() {
 
   const pathElem = document.createElementNS(xmlns, 'path')
   pathElem.setAttribute('role', 'presentation')
-  pathElem.setAttribute('d', p[0])
+  pathElem.setAttribute('d', d[0])
   pathElem.textContent = 'No newline at end of file'
   svgElem.appendChild(pathElem)
