+++ app/src/ui/lib/avatar.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { IAvatarUser } from '../../models/avatar'
-import { Octicon } from '../octicons'
+import { CustomOcticonSymbolType, Octicon } from '../octicons'
 import { API, getDotComAPIEndpoint, getHTMLURL } from '../../lib/api'
 import { TooltippedContent } from './tooltipped-content'
 import { TooltipDirection } from './tooltip'
@@ -182,11 +182,11 @@ interface IAvatarState {
  * The octicon has been tweaked to add some padding and so that it scales nicely in
  * a square aspect ratio.
  */
-const DefaultAvatarSymbol = {
+const DefaultAvatarSymbol: CustomOcticonSymbolType = {
   s: 'default-avatar',
   w: 16,
   h: 16,
-  p: [
+  d: [
     'M13 13.145a.844.844 0 0 1-.832.855H3.834A.846.846 0 0 1 3 13.142v-.856c0-2.257 3.333-3.429 3.333-3.429s.191-.35 0-.857c-.7-.531-.786-1.363-.833-3.429C5.644 2.503 7.056 2 8 2s2.356.502 2.5 2.571C10.453 6.637 10.367 7.47 9.667 8c-.191.506 0 .857 0 .857S13 10.03 13 12.286v.859z',
   ],
 }
