+++ app/src/ui/octicons/sync-clockwise.ts
@@ -1,4 +1,4 @@
-import { CustomOcticonSymbolType } from './octicons.generated'
+import { CustomOcticonSymbolType } from '.'
 
 /**
  * An horizontally flipped (i.e `scaleX(-1)`) but otherwise unmodified version of
@@ -8,7 +8,7 @@ export const syncClockwise: CustomOcticonSymbolType = {
   s: 'sync-clockwise',
   w: 16,
   h: 16,
-  p: [
+  d: [
     'M8 2.5c1.645 0 3.123.722 4.131 1.869l-1.204 1.204a.25.25 0 0 0 .177.427h3.646a.25.25 ' +
       '0 0 0 .25-.25V2.104a.25.25 0 0 0-.427-.177l-1.38 1.38A7.001 7.001 0 0 0 1.05 7.16a.75.75 ' +
       '0 1 0 1.49.178A5.501 5.501 0 0 1 8 2.5zm6.294 5.505a.75.75 0 0 0-.833.656 5.501 5.501 ' +
